requirements.txt
Django==5.1.7
djangorestframework==3.15.2
flake8==7.1.2
python-dotenv==1.0.1
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
whitenoise==6.9.0
